+++ ost/graphs/ard_json/slc.ost_gtc.json
@@ -1,5 +1,5 @@
 {
-    "processing_parameters": {
+    "processing": {
         "single_ARD": {
             "type": "OST-GTC",
             "resolution": 20,
